+++ lib/src/headers.dart
@@ -36,6 +36,12 @@ abstract class ZoomableHeadersWidget<C extends ZoomController> extends StatefulW
   /// Whether the widget should automatically be placed in a scrollable widget.
   final bool inScrollableWidget;
 
+  /// The initial visible hour.
+  final int initialHour;
+
+  /// The initial visible minute.
+  final int initialMinute;
+
   /// Whether the widget should automatically scroll to the current time (hour and minute).
   final bool scrollToCurrentTime;
 
@@ -58,6 +64,8 @@ abstract class ZoomableHeadersWidget<C extends ZoomController> extends StatefulW
     this.hoursColumnBackgroundColor,
     double hourRowHeight = 60,
     @required this.inScrollableWidget,
+    int initialHour,
+    int initialMinute,
     @required this.scrollToCurrentTime,
     @required this.userZoomable,
   })  : assert(dateFormatter != null),
@@ -65,6 +73,8 @@ abstract class ZoomableHeadersWidget<C extends ZoomController> extends StatefulW
         dayBarHeight = math.max(0, dayBarHeight ?? 40),
         hoursColumnWidth = math.max(0, hoursColumnWidth ?? 60),
         hourRowHeight = math.max(0, hourRowHeight ?? 60),
+        initialHour = math.min(23, math.max(0, initialHour ?? 0)),
+        initialMinute = math.min(59, math.max(0, initialMinute ?? 0)),
         assert(inScrollableWidget != null),
         assert(scrollToCurrentTime != null),
         assert(userZoomable != null);
@@ -83,12 +93,6 @@ abstract class ZoomableHeadersWidgetState<W extends ZoomableHeadersWidget<C>, C 
     super.initState();
     hourRowHeight = widget._calculateHourRowHeight();
     widget.controller.addListener(this);
-
-    if (shouldScrollToCurrentTime) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        scrollToCurrentTime();
-      });
-    }
   }
 
   @override
@@ -134,25 +138,39 @@ abstract class ZoomableHeadersWidgetState<W extends ZoomableHeadersWidget<C>, C 
     super.dispose();
   }
 
-  /// Schedules a scroll to the current time if needed.
-  void scheduleScrollToCurrentTimeIfNeeded() {
-    if (shouldScrollToCurrentTime) {
-      WidgetsBinding.instance.addPostFrameCallback((_) => scrollToCurrentTime());
+  /// Schedules a scroll to the default hour.
+  void scheduleScrollToInitialHour() {
+    if (mounted) {
+      WidgetsBinding.instance.addPostFrameCallback((_) => scrollToTime(widget.initialHour, widget.initialMinute));
     }
   }
 
+  /// Schedules a scroll to the current time if needed.
+  bool scheduleScrollToCurrentTimeIfNeeded() {
+    if (shouldScrollToCurrentTime) {
+      WidgetsBinding.instance.addPostFrameCallback((_) => scrollToCurrentTime());
+      return true;
+    }
+    return false;
+  }
+
   /// Checks whether the widget should scroll to current time.
-  bool get shouldScrollToCurrentTime => widget.scrollToCurrentTime && widget.inScrollableWidget;
+  bool get shouldScrollToCurrentTime => widget.scrollToCurrentTime;
 
   /// Scrolls to current time.
   void scrollToCurrentTime() {
-    if (!mounted) {
-      return;
+    if (mounted) {
+      DateTime now = DateTime.now();
+      scrollToTime(now.hour, now.minute);
     }
+  }
 
-    DateTime now = DateTime.now();
-    double topOffset = calculateTopOffset(now.hour, now.minute);
-    widget.controller.verticalScrollController.jumpTo(math.min(topOffset, widget.controller.verticalScrollController.position.maxScrollExtent));
+  /// Scrolls to a given time if possible.
+  void scrollToTime(int hour, int minute) {
+    if (widget.inScrollableWidget) {
+      double topOffset = calculateTopOffset(hour, minute);
+      widget.controller.verticalScrollController.jumpTo(math.min(topOffset, widget.controller.verticalScrollController.position.maxScrollExtent));
+    }
   }
 
   /// Returns whether this widget should be zoomable.